refactor(phase4): extract requirements and args utilities from bin/cli.js
Part of #89 - Phase 4: Extract requirements & args utilities

Extracted utility functions:
- parseArgs() → lib/utils/args-parser.js (18 lines)
- usage() → lib/utils/cli-help.js (16 lines)
- resolvePkgVersion() + gte() → lib/utils/version-utils.js (22 lines)
- checkRequirements() → lib/utils/requirements.js (38 lines)

Progress:
- bin/cli.js: 473 → 407 lines (-66 lines)
- Total progress: 769 → 407 lines (-362 lines, -47%)
- All 554 tests passing

Fourth of 6 PRs to refactor bin/cli.js to a thin router (~40 lines)
