+++ lib/utils/args-parser.js
@@ -1,0 +1,18 @@
+'use strict';
+
+function parseArgs(argv) {
+  // simple parse; supports flags like --foo or --foo=bar
+  const out = { _: [] };
+  for (let i = 2; i < argv.length; i++) {
+    const a = argv[i];
+    if (a.startsWith('--')) {
+      const [k, v] = a.replace(/^--/, '').split('=');
+      out[k] = v === undefined ? true : v;
+    } else {
+      out._.push(a);
+    }
+  }
+  return out;
+}
+
+module.exports = { parseArgs };
