+++ lib/utils/version-utils.js
@@ -1,0 +1,22 @@
+'use strict';
+
+function resolvePkgVersion(name, cwd) {
+  try {
+    const pkg = require(require.resolve(`${name}/package.json`, { paths: [cwd] }));
+    return pkg.version || null;
+  } catch {
+    return null;
+  }
+}
+
+function gte(a, b) {
+  const pa = String(a).split('.').map(Number);
+  const pb = String(b).split('.').map(Number);
+  for (let i = 0; i < 3; i++) {
+    if ((pa[i] || 0) > (pb[i] || 0)) return true;
+    if ((pa[i] || 0) < (pb[i] || 0)) return false;
+  }
+  return true;
+}
+
+module.exports = { resolvePkgVersion, gte };
